+++ packages/twenty-server/src/modules/dashboard-sync/dashboard-sync.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 
-import { DashboardSyncService } from 'src/engine/metadata-modules/dashboard-sync/services/dashboard-sync.service';
 import { WorkspaceManyOrAllFlatEntityMapsCacheModule } from 'src/engine/metadata-modules/flat-entity/services/workspace-many-or-all-flat-entity-maps-cache.module';
 import { TwentyORMModule } from 'src/engine/twenty-orm/twenty-orm.module';
+import { DashboardSyncService } from 'src/modules/dashboard-sync/services/dashboard-sync.service';
 
 @Module({
   imports: [TwentyORMModule, WorkspaceManyOrAllFlatEntityMapsCacheModule],
